fix jshooks build error
