+++ src/test/jtx/hook.h
@@ -20,6 +20,7 @@
 #ifndef RIPPLE_TEST_JTX_HOOK_H_INCLUDED
 #define RIPPLE_TEST_JTX_HOOK_H_INCLUDED
 
+#include <ripple/protocol/jss.h>
 #include <ripple/json/json_value.h>
 #include <optional>
 #include <test/jtx/Account.h>
@@ -46,6 +47,29 @@ hso(std::string const& wasmHex, void (*f)(Json::Value& jv) = 0);
 
 Json::Value
 hso_delete(void (*f)(Json::Value& jv) = 0);
+
+class JSSHasher
+{
+public:
+    size_t
+    operator()(const Json::StaticString& n) const
+    {
+        return std::hash<std::string_view>{}(n.c_str());
+    }
+};
+
+class JSSEq
+{
+public:
+    bool
+    operator()(const Json::StaticString& a, const Json::StaticString& b) const
+    {
+        return a == b;
+    }
+};
+
+using JSSMap =
+    std::unordered_map<Json::StaticString, Json::Value, JSSHasher, JSSEq>;
 
 }  // namespace jtx
 }  // namespace test
